fix #26, compile error on Lolin Wemos D1

+++ examples/SHT85_demo_async/SHT85_demo_async.ino
@@ -18,7 +18,7 @@
 
 #include "SHT85.h"
 
-#define SHT85_ADDRESS         0x44
+#define SHT85_ADDRESS         ((uint8_t)0x44)
 
 SHT85 sht(SHT85_ADDRESS);
